+++ website/builder/src/.vuepress/config.ts
@@ -26,6 +26,10 @@ function getNavBar(lang: 'cn' | 'en'): NavItem[] {
       text: 'API',
       link: getLink('/api/'),
     },
+    {
+      text: getText('插件', 'Plugins'),
+      link: getLink('/plugins/introduction'),
+    },
   ];
 }
 
@@ -35,7 +39,7 @@ function getSidebar(lang: 'cn' | 'en'): SidebarConfig4Multiple {
     [getLink('/guide/')]: [
       {
         collapsable: false,
-        title: getText('开始', 'Start'),
+        title: getText('开始', 'Getting Started'),
         children: [
           getLink('/guide/introduction'),
           getLink('/guide/quick-start'),
@@ -46,16 +50,12 @@ function getSidebar(lang: 'cn' | 'en'): SidebarConfig4Multiple {
       {
         collapsable: false,
         title: getText('基础', 'Basic'),
-        children: [
-          getLink('/guide/basic/build-target'),
-        ],
+        children: [getLink('/guide/basic/build-target')],
       },
       {
         collapsable: false,
         title: getText('进阶', 'Advanced'),
-        children: [
-          getLink('/guide/advanced/build-performance'),
-        ],
+        children: [getLink('/guide/advanced/build-performance')],
       },
       {
         collapsable: false,
@@ -92,13 +92,36 @@ function getSidebar(lang: 'cn' | 'en'): SidebarConfig4Multiple {
           getLink('/api/builder-core'),
           getLink('/api/builder-instance'),
           getLink('/api/builder-types'),
+          getLink('/api/plugin-core'),
+          getLink('/api/plugin-hooks'),
+        ],
+      },
+    ],
+    [getLink('/plugins')]: [
+      {
+        collapsable: false,
+        title: getText('指南', 'Guide'),
+        children: [
+          getLink('/plugins/introduction'),
+          getLink('/plugins/integration'),
+          getLink('/plugins/testing'),
+        ],
+      },
+      {
+        collapsable: false,
+        title: getText('列表', 'List'),
+        children: [
+          getLink('/plugins/list'),
+          getLink('/plugins/plugin-swc'),
+          getLink('/plugins/plugin-esbuild'),
+          getLink('/plugins/plugin-node-polyfill'),
         ],
       },
     ],
   };
 }
 
-export default defineConfig4CustomTheme<ThemeConfig>((ctx) => ({
+export default defineConfig4CustomTheme<ThemeConfig>(ctx => ({
   base: '/builder/',
   head: [
     ['link', { rel: 'icon', href: `https://modernjs.dev/img/favicon.ico` }],
@@ -175,10 +198,10 @@ export default defineConfig4CustomTheme<ThemeConfig>((ctx) => ({
     // OptimizeCssAssetsWebpackPlugin will cause the build to fail,
     // removed will not affect the build result
     config.plugins = config.plugins.filter(
-      (plugin) => plugin.constructor.name !== 'OptimizeCssAssetsWebpackPlugin'
+      plugin => plugin.constructor.name !== 'OptimizeCssAssetsWebpackPlugin',
     );
   },
-  extendMarkdown: (md) => {
+  extendMarkdown: md => {
     md.use(markdownItInclude);
   },
 }));