+++ packages/generator/generator-utils/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @modern-js/generator-utils
 
+## 2.4.7
+
+### Patch Changes
+
+- Updated dependencies [7c89bd6]
+  - @modern-js/generator-common@2.4.7
+  - @modern-js/plugin-i18n@1.21.2
+  - @modern-js/utils@1.21.2
+
 ## 2.4.6
 
 ### Patch Changes