+++ packages/generator/generators/repo-generator/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @modern-js/repo-generator
 
+## 2.4.7
+
 ## 2.4.6
 
 ## 2.4.5